- combigrid update, the correct combination of FGs into an SG
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2362 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/combigrid/combigrid/AdaptiveSerialCombiGrid.cpp
@@ -10,7 +10,7 @@
 std::vector<int> combigrid::AdaptiveSerialCombiGrid::addToCombiScheme(
 		std::vector<int> level) {
 
-	COMBIGRID_ERROR_TEST(level.size()==abs(combikernel_->getDim()),
+	COMBIGRID_ERROR_TEST((int)level.size()==abs(combikernel_->getDim()),
 			"new level vector has the wrong size!")
 	CombigridLevelVector current(combischeme_->getLevels(),
 			combischeme_->getCoef());
